+++ Commun/Utilitaire/Vue/ProjectionOrtho.h
@@ -132,7 +132,7 @@ namespace vue {
 
 	////////////////////////////////////////////////////////////////////////
 	///
-	/// @fn inline glm::ivec2 ProjectionOrtho::obtenirDimensionFenetreVirtuelle() const
+	/// @fn inline glm::dvec2 ProjectionOrtho::obtenirDimensionFenetreVirtuelle() const
 	///
 	/// Cette fonction retourne les dimensions de la fenêtre virtuelle
 	/// associée à cette projection orthographique.
@@ -142,7 +142,7 @@ namespace vue {
 	////////////////////////////////////////////////////////////////////////
 	inline glm::dvec2 ProjectionOrtho::obtenirDimensionFenetreVirtuelle() const
 	{
-		return glm::ivec2{ xMaxFenetre_ - xMinFenetre_, yMaxFenetre_ - yMinFenetre_ };
+		return glm::dvec2{ xMaxFenetre_ - xMinFenetre_, yMaxFenetre_ - yMinFenetre_ };
 	}
 }; // Fin de l'espace de nom vue.
